@@ -54,24 +62,23 @@
     "ajv-keywords": "^3.4.0",
     "boron": "^0.2.3",
     "brace": "^0.10.0",
-    "classnames": "^2.1.3",
-    "core-js": "^2.4.1",
-    "deepmerge": "^1.3.2",
-    "file-dialog": "0.0.7",
+    "classnames": "^2.2.6",
+    "deepmerge": "^4.0.0",
+    "file-dialog": "0.0.8",
     "immutable": "^3.x.x",
     "is-json": "^2.0.1",
     "js-yaml": "^3.13.1",
     "json-beautify": "^1.0.1",
     "json-refs": "^3.0.4",
-    "lodash": "^4.17.15",
-    "promise-worker": "^1.1.1",
-    "prop-types": "15.6.0",
+    "lodash": "=4.17.21",
+    "promise-worker": "^2.0.1",
+    "prop-types": "15.7.2",
     "querystring-browser": "^1.0.4",
-    "react": "^15.6.2",
+    "react": "=15.7.0",
     "react-ace": "^4.1.6",
     "react-addons-css-transition-group": "^15.4.2",
     "react-dd-menu": "^2.0.0",
-    "react-dom": "^15.6.2",
+    "react-dom": "=15.7.0",
     "react-dropzone": "4.2.11",
     "react-file-download": "^0.3.2",
     "react-immutable-proptypes": "^2.1.0",
@@ -79,89 +86,85 @@
     "react-split-pane": "^0.1.82",
     "react-transition-group": "^1.1.1",
     "redux": "^3.x.x",
-    "reselect": "^2.5.4",
-    "swagger-client": "^3.9.5",
-    "swagger-ui": "^3.24.2",
+    "reselect": "^4.0.0",
+    "swagger-client": "^3.13.1",
+    "swagger-ui": "^3.43.0",
     "traverse": "^0.6.6",
-    "whatwg-fetch": "^2.0.3",
-    "yaml-js": "^0.1.4"
+    "yaml-js": "^0.2.3"
   },
   "devDependencies": {
-    "@babel/cli": "^7.5.5",
-    "@babel/core": "^7.0.0",
-    "@babel/plugin-proposal-class-properties": "^7.5.5",
-    "@babel/plugin-proposal-optional-chaining": "^7.2.0",
-    "@babel/plugin-transform-runtime": "^7.5.5",
-    "@babel/preset-env": "^7.0.0",
-    "@babel/preset-react": "^7.0.0",
-    "@babel/register": "^7.5.5",
-    "@babel/runtime-corejs2": "^7.0.0",
-    "@release-it/conventional-changelog": "^1.1.0",
-    "autoprefixer": "^6.6.1",
+    "@babel/cli": "=7.13.0",
+    "@babel/core": "=7.13.1",
+    "@babel/plugin-proposal-class-properties": "=7.13.0",
+    "@babel/plugin-proposal-optional-chaining": "=7.13.0",
+    "@babel/plugin-transform-runtime": "=7.13.6",
+    "@babel/preset-env": "=7.13.5",
+    "@babel/preset-react": "=7.12.13",
+    "@babel/register": "=7.13.0",
+    "@babel/runtime-corejs3": "^7.11.2",
+    "@commitlint/cli": "^11.0.0",
+    "@commitlint/config-conventional": "^11.0.0",
+    "@jest/globals": "=26.6.2",
+    "@release-it/conventional-changelog": "=1.1.4",
+    "autoprefixer": "=9.8.6",
     "babel-eslint": "^10.0.2",
-    "babel-loader": "^8.0.6",
-    "babel-plugin-module-alias": "^1.6.0",
-    "babel-plugin-module-resolver": "^3.2.0",
+    "babel-loader": "^8.1.0",
+    "babel-plugin-module-resolver": "^4.0.0",
     "babel-plugin-transform-es2015-constants": "^6.1.4",
     "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
-    "bundlesize": "^0.17.2",
-    "cross-env": "^5.1.3",
-    "css-loader": "^3.1.0",
+    "cross-env": "=7.0.3",
+    "css-loader": "^5.0.2",
     "cssnano": "^4.1.10",
-    "cypress": "^3.4.1",
+    "cypress": "=5.3.0",
+    "dedent": "^0.7.0",
     "deep-extend": "^0.6.0",
-    "deepmerge": "^1.3.2",
     "enzyme": "^3.3.0",
     "enzyme-adapter-react-15": "^1.0.5",
-    "eslint": "^6.1.0",
-    "eslint-plugin-import": "^2.13.0",
-    "eslint-plugin-react": "^4.3.0",
+    "eslint": "^7.18.0",
+    "eslint-plugin-import": "^2.21.1",
+    "eslint-plugin-jest": "=23.20.0",
+    "eslint-plugin-mocha": "^8.0.0",
+    "eslint-plugin-react": "^7.20.0",
     "expect": "^1.20.2",
-    "extract-text-webpack-plugin": "0.8.2",
-    "file-loader": "0.8.4",
+    "file-loader": "6.2.0",
     "git-describe": "^4.0.1",
-    "html-webpack-plugin": "^2.28.0",
-    "http-server": "^0.11.1",
+    "http-server": "^0.12.3",
+    "husky": "=4.3.8",
     "ignore-assets-webpack-plugin": "^2.0.1",
-    "jsdom": "11.12.0",
+    "inspectpack": "=4.6.1",
+    "jest": "=26.6.3",
+    "jsdom": "=16.4.0",
     "jsdom-global": "3.0.2",
-    "json-loader": "0.5.3",
-    "less": "^3.9.0",
-    "less-loader": "^5.0.0",
-    "license-checker": "^8.0.4",
-    "matcher": "^0.1.2",
-    "mini-css-extract-plugin": "^0.8.0",
-    "mocha": "^6.2.0",
-    "mocha-jsdom": "^2.0.0",
-    "npm-audit-ci-wrapper": "^2.3.0",
-    "npm-run-all": "^4.1.3",
-    "null-loader": "0.1.1",
-    "open": "^6.4.0",
+    "json-loader": "0.5.7",
+    "less": "^3.11.2",
+    "less-loader": "^7.2.1",
+    "license-checker": "^25.0.0",
+    "lint-staged": "=10.5.4",
+    "mini-css-extract-plugin": "=1.3.8",
+    "mocha": "^8.2.1",
+    "npm-audit-ci-wrapper": "^3.0.1",
+    "npm-run-all": "^4.1.5",
+    "null-loader": "4.0.1",
+    "open": "^7.0.4",
     "postcss-loader": "^3.0.0",
-    "raw-loader": "0.5.1",
+    "postcss-preset-env": "=6.7.0",
+    "raw-loader": "3.1.0",
     "react-hot-loader": "^1.3.1",
     "react-test-renderer": "^15.6.2",
-    "release-it": "^12.3.5",
+    "release-it": "=13.7.1",
     "rewiremock": "^3.7.4",
-    "rimraf": "^2.6.0",
-    "standard": "^8.6.0",
-    "style-loader": "0.13.0",
-    "uglifyjs-webpack-plugin": "^1.2.5",
-    "url-loader": "^2.1.0",
-    "webpack": "^4.36.1",
-    "webpack-bundle-size-analyzer": "^2.5.0",
-    "webpack-cli": "^3.3.6",
-    "webpack-dev-server": "^3.7.2",
+    "rimraf": "^3.0.2",
+    "style-loader": "^1.3.0",
+    "terser-webpack-plugin": "^4.2.3",
+    "url-loader": "^4.1.1",
+    "webpack": "^4.43.0",
+    "webpack-bundle-size-analyzer": "=3.1.0",
+    "webpack-cli": "^3.3.11",
+    "webpack-dev-server": "^3.11.0",
+    "webpack-stats-plugin": "=1.0.3",
     "worker-loader": "^2.0.0"
   },
   "config": {
     "deps_check_dir": ".deps_check"
-  },
-  "bundlesize": [
-    {
-      "path": "./dist/swagger-editor-bundle.js",
-      "maxSize": "2.8 MB",
-      "compression": "none"
-    }
-  ]
+  }
 }
