@@ -194,9 +194,9 @@ export function pathForPosition(yaml, position) {
 
     const problemMark = e.problem_mark || {}
     const errorTraceMessage = [
-      yaml.split("\n").slice(problemMark.line - 5, problemMark.line).join("\n"),
+      yaml.split("\n").slice(problemMark.line - 5, problemMark.line + 1).join("\n"),
       Array(problemMark.column).fill(" ").join("") + `^----- ${e.name}: ${e.toString().split("\n")[0]}`,
-      yaml.split("\n").slice(problemMark.line, problemMark.line + 5).join("\n")
+      yaml.split("\n").slice(problemMark.line + 1, problemMark.line + 5).join("\n")
     ].join("\n")
 
     console.error(errorTraceMessage)
