@@ -16,6 +16,8 @@ export const isOAS3RootRequestBody = (state, node) => node.path.length === 3 && 
 export const isOAS3OperationRequestBody = (state, node) => node.path.length === 4 && node.path[3] === "requestBody"
 export const isOAS3OperationCallbackRequestBody = (state, node) => node.path.length === 8 && node.path[7] === "requestBody"
 export const isOAS3RootParameter = (state, node) => node.path[0] === "components" && node.path[1] === "parameters" && node.path.length === 3
+export const isOAS3RootResponse = (state, node) => node.path[0] === "components" && node.path[1] === "responses" && node.path.length === 3
+export const isOAS3RootSchema = (state, node) => node.path[0] === "components" && node.path[1] === "schemas" && node.path.length === 3
 
 export const isSubSchema = (state, node) => (sys) => {
   const path = node.path
@@ -94,15 +96,29 @@ export const isOAS3ResponseSchema = (state, node) => () => {
 }
 
 export const isResponse = (state, node) => (sys) => {
-  if(sys.validateSelectors.isVendorExt(node)) {
-    return false
-  }
-  return (
-    sys.validateSelectors.isRootResponse(node)
-      || ( node.path[0] === "paths"
-           && node.path[3] === "responses"
-           && node.path.length === 5)
+  const isOperationResponse = (
+    node.path[0] === "paths"
+      && node.path[3] === "responses"
+      && node.path.length === 5
+      && !sys.validateSelectors.isVendorExt(node)
   )
+
+  return (
+    isOperationResponse
+      || sys.validateSelectors.isRootResponse(node)
+      || sys.validateSelectors.isOAS3RootResponse(node)
+  )
+}
+
+export const allResponses = () => (system) => {
+  return system.fn.traverseOnce({
+    name: "allResponses",
+    fn: (node) => {
+      if(system.validateSelectors.isResponse(node)) {
+        return node
+      }
+    },
+  })
 }
 
 export const isHeader = (state, node) => (sys) => {
@@ -224,7 +240,10 @@ export const allDefinitions = () => (system) => {
   return system.fn.traverseOnce({
     name: "allDefinitions",
     fn: (node) => {
-      if(system.validateSelectors.isDefinition(node)) {
+      if(
+        system.validateSelectors.isDefinition(node)
+         || system.validateSelectors.isOAS3RootSchema(node)
+       ) {
         return node
       }
     },
@@ -249,6 +268,19 @@ export const allOAS3OperationSchemas = () => (system) => {
       if(
         system.validateSelectors.isOAS3RequestBodySchema(node)
          || system.validateSelectors.isOAS3ResponseSchema(node)
+       ) {
+        return node
+      }
+    },
+  })
+}
+
+export const allOAS3RequestBodySchemas = () => (system) => {
+  return system.fn.traverseOnce({
+    name: "allOAS3RequestBodySchemas",
+    fn: (node) => {
+      if(
+        system.validateSelectors.isOAS3RequestBodySchema(node)
        ) {
         return node
       }
@@ -319,10 +351,15 @@ export const allSecurityDefinitions = () => (system) => {
       const isSecurityDefinition = (
         node.path[0] == "securityDefinitions"
           && node.path.length === 2
-          && !system.validateSelectors.isVendorExt(node)
       )
 
-      if(isSecurityDefinition) {
+      const isOAS3SecurityScheme = (
+        node.path[0] == "components"
+          && node.path[1] == "securitySchemes"
+          && node.path.length === 3
+      )
+
+      if(isSecurityDefinition || isOAS3SecurityScheme) {
         return node
       }
     }
@@ -336,17 +373,34 @@ export const allSecurityRequirements = () => (system) => {
       const isGlobalSecurityRequirement = (
         node.path[0] == "security"
           && node.path.length === 2
-          && !system.validateSelectors.isVendorExt(node)
       )
 
       const isOperationSecurityRequirement = (
         node.path[0] == "paths"
           && node.path[3] == "security"
           && node.path.length === 5
-          && !system.validateSelectors.isVendorExt(node)
+          && !system.validateSelectors.isVendorExt(node.parent) // ignore extension keys in path items
+          && !system.validateSelectors.isVendorExt(node.parent.parent.parent) // ignore extension keys in "paths"
       )
 
       if(isGlobalSecurityRequirement || isOperationSecurityRequirement) {
+        return node
+      }
+    }
+  })
+}
+
+export const allOAS3Components = () => (system) => {
+  return system.fn.traverseOnce({
+    name: "allOAS3Components",
+    fn: (node) => {
+      const isComponent = (
+        node.path[0] === "components"
+          && node.path.length === 3
+          && !system.validateSelectors.isVendorExt(node.parent)
+      )
+
+      if(isComponent) {
         return node
       }
     }
