@@ -33,42 +33,42 @@ export const validate2And3ParametersHaveUniqueNameAndInCombinations = () => (sys
     })
 }
 
-export const validate2And3ParameterDefaultsMatchAnEnum = () => (system) => {
+export const validate2And3PathParameterIsDefinedInPath = () => (system) => {
+  const refArray = []
   return system.validateSelectors
     .allParameters()
     .then(nodes => {
       return nodes.reduce((acc, node) => {
         const parameter = node.node || {}
-        const isOAS3 = system.specSelectors.isOAS3()
-        let paramEnum, paramDefault, internalLocation
-
-
-        if(isOAS3) {
-          const schema = parameter.schema
-          if(!schema || schema.enum === undefined || schema.default === undefined) {
-            // nothing to do
-            return acc
+        const path = node.path
+        const isFromPath = path[0] === "paths" ? true : false
+        const pathString = path[1]
+        const paramName = parameter.name
+        const paramInPath = `{${paramName}}`
+        const ref = parameter.$ref
+        const pathStringIncludesParamInPath = pathString && !pathString.toUpperCase().includes("" + paramInPath.toUpperCase())
+        if (parameter.in === "path") {
+          if (isFromPath && pathStringIncludesParamInPath) {
+            acc.push({
+              message: `Path parameter "${paramName}" must have the corresponding ${paramInPath} segment in the "${pathString}" path`,
+              path: [...node.path, "name"],
+              level: "error"
+            })
+          } else {
+            const paramReference = refArray.find(({ referenceParamName }) => referenceParamName === node.key) 
+            if (paramReference && paramReference.pathString && !paramReference.pathString.toUpperCase().includes("" + paramInPath.toUpperCase())) {
+              acc.push({
+                message: `Path parameter "${paramName}" must have the corresponding ${paramInPath} segment in the "${paramReference.pathString}" path`,
+                path: [...paramReference.node.path, "name"],
+                level: "error"
+              })
+            }
           }
-          paramEnum = schema.enum
-          paramDefault = schema.default
-          internalLocation = ["schema", "default"]
-        } else {
-          if(!parameter || parameter.enum === undefined || parameter.default === undefined) {
-            // nothing to do
-            return acc
-          }
-          paramEnum = parameter.enum
-          paramDefault = parameter.default
-          internalLocation = ["default"]
+        } else if (ref !== undefined) {
+          const refStrings = ref.split("/")
+          refArray.push({referenceParamName:refStrings[refStrings.length-1], pathString:pathString, node: node})
         }
-
-        if(paramEnum.indexOf(paramDefault) === -1) {
-          acc.push({
-            message: "Default values must be present in `enum`",
-            path: [...node.path, ...internalLocation]
-          })
-        }
-
+        
         return acc
       }, [])
     })
