@@ -12,14 +12,21 @@ class FormInput extends Component {
   render() { 
     return (
       <div>
-        <input 
-          type="text"
-          value={this.props.inputValue}
-          className={`form-control ${this.props.isValid || this.isNotRequiredAndEmpty() ? "" : "border border-danger"}`} 
-          onChange={this.props.onChange}
-          placeholder={this.props.placeholderText}
-        />
-
+        {this.props.bigTextBox ? 
+          <textarea
+            type="text"
+            value={this.props.inputValue}
+            className={`form-control ${this.props.isValid || this.isNotRequiredAndEmpty() ? "" : "border border-danger"}`} 
+            onChange={this.props.onChange}
+            placeholder={this.props.placeholderText}
+          />:
+          <input 
+            type="text"
+            value={this.props.inputValue}
+            className={`form-control ${this.props.isValid || this.isNotRequiredAndEmpty() ? "" : "border border-danger"}`} 
+            onChange={this.props.onChange}
+            placeholder={this.props.placeholderText}
+          />}
         {!this.props.isValid && !this.isNotRequiredAndEmpty() && this.props.validationMessage && 
           <div className="invalid-feedback">
             {this.props.validationMessage}
@@ -36,7 +43,8 @@ FormInput.propTypes = {
   validationMessage: PropTypes.string,
   inputValue: PropTypes.string.isRequired,
   onChange: PropTypes.func.isRequired,
-  isRequired: PropTypes.bool
+  isRequired: PropTypes.bool,
+  bigTextBox: PropTypes.bool
 }
 
 export default FormInput 
