@@ -8,6 +8,8 @@ import { tagsForm, tagsObject } from "./forms/form-objects/tags-object"
 import { serversForm, serversObject } from "./forms/form-objects/servers-object"
 import { externalDocumentationForm, externalDocumentationObject } from "./forms/form-objects/external-documentation-object"
 import { addOperationTagsForm, addOperationTagsObject } from "./forms/form-objects/add-operation-tags"
+import { exampleForm, exampleObject } from "./forms/form-objects/example-value-object"
+import { selectOperationObject } from "./forms/form-objects/select-operation"
 
 export default class TopbarInsert extends Component {
   constructor(props) {
@@ -33,6 +35,9 @@ export default class TopbarInsert extends Component {
     this.updateServers = this.updateServers.bind(this)
     this.updateTags = this.updateTags.bind(this)
     this.addOperationTags = this.addOperationTags.bind(this)
+    this.getResponses = this.getResponses.bind(this)
+    this.getMediaTypes = this.getMediaTypes.bind(this)
+    this.addExampleResponse = this.addExampleResponse.bind(this)
   }
 
   openModalClick = showModalProperty => () => {
@@ -60,7 +65,7 @@ export default class TopbarInsert extends Component {
     this.props.specActions.addToSpec([], infoObject(formData), "info")
   } 
 
-  getPaths = () => this.props.specSelectors.paths() ? Object.keys(this.props.specSelectors.paths().toJS()) : null;
+  getPaths = () => this.props.specSelectors.paths() ? Object.keys(this.props.specSelectors.paths().toJS()) : null
 
   addOperationTags(formData) {
     const operationTagsObject = addOperationTagsObject(formData)
@@ -91,6 +96,54 @@ export default class TopbarInsert extends Component {
 
   updateTags = (formData) => {
     this.props.specActions.addToSpec(["tags"], tagsObject(formData), null)
+  }
+
+  getResponses = (method, formData) => {
+    // Operation = 'depends on' value
+    // formData = all the data so far, process to get the response as well
+    const operationPath = [...selectOperationObject(formData), "responses"]
+    const responses = this.props.specSelectors.specJson().getIn(operationPath)
+
+    if (!responses) {
+      return []
+    }
+
+    return Object.keys(responses.toJS())
+  }
+
+  getMediaTypes = (response, formData) => {
+    const defaultOptions = [
+      "application/json", 
+      "text/plain; charset=utf-8", 
+      "application/xml" ]
+
+    if (!formData) {  
+      return defaultOptions
+    }
+
+    // Operation = 'depends on' value
+    // formData = all the data so far, process to get the response as well
+    const operationPath = [...selectOperationObject(formData), "responses"]
+    const responses = this.props.specSelectors.specJson().getIn(operationPath)
+
+    if (responses) {
+      const response = responses.get(formData.getIn(["response", "value"]))
+
+      if (response && response.has("content")) {
+        const existing = Object.keys(response.get("content").toJS())
+        const combined = defaultOptions.concat(existing)
+
+        // Remove duplicates.
+        return combined.filter((item, pos) => combined.indexOf(item) == pos)
+      }
+    }
+
+    return defaultOptions
+  }
+
+  addExampleResponse = (formData) => {
+    const formObject = exampleObject(formData)
+    this.props.specActions.addToSpec(formObject.responsePath, { value: formObject.exampleValue }, formObject.exampleName)
   }
 
   render() {
@@ -193,7 +246,6 @@ export default class TopbarInsert extends Component {
             />
           </Modal>
         }
-        
         { this.state.showAddOperationTagsModal && 
           <Modal
             title="Add Tags To Operation"
@@ -211,6 +263,23 @@ export default class TopbarInsert extends Component {
             />
           </Modal>
         }
+        { this.state.showAddExampleModal && 
+          <Modal
+            title="Add Example Response"
+            description="An example response sent from the API."
+            onCloseClick={this.closeModalClick("showAddExampleModal")}
+            isShownisLarge
+          >
+            <AddForm
+              {...this.props}
+              submit={this.closeModalClick("showAddExampleModal")}
+              getFormData={exampleForm}
+              existingData={{ getPaths: this.getPaths, getOperations: this.getOperations, getResponses: this.getResponses, getMediaTypes: this.getMediaTypes }}
+              submitButtonText="Add Example Response"
+              updateSpecJson={this.addExampleResponse}
+            />
+          </Modal>
+        }
 
         <Dropdown displayName="Insert" >
           <DropdownItem onClick={this.openModalClick("showAddPathModal")} name="Add Path Item"/>    
@@ -220,6 +289,7 @@ export default class TopbarInsert extends Component {
           <DropdownItem onClick={this.openModalClick("showAddTagsModal")} name="Add Tag Declarations" />
           <DropdownItem onClick={this.openModalClick("showAddOperationTagsModal")} name="Add Tags To Operation" />
           <DropdownItem onClick={this.openModalClick("showAddServersModal")} name="Add Servers" />
+          <DropdownItem onClick={this.openModalClick("showAddExampleModal")} name="Add Example Response" />
         </Dropdown>
       </div>
     )
@@ -229,7 +299,7 @@ export default class TopbarInsert extends Component {
 TopbarInsert.propTypes = {
   getComponent: PropTypes.func.isRequired,
   specSelectors: PropTypes.shape({
-    specStr: PropTypes.func.isRequired,
+    specJson: PropTypes.func.isRequired,
     info: PropTypes.func.isRequired,
     paths: PropTypes.func.isRequired,
     isOAS3: PropTypes.func.isRequired,
