@@ -1,43 +1,42 @@
-import SwaggerUi from "swagger-ui"
-import EditorSpecPlugin from "src/plugins/editor/spec"
-import expect from "expect"
+import SwaggerUi from 'swagger-ui';
+import EditorSpecPlugin from 'src/plugins/editor/spec';
 
-describe("editor state plugins", function() {
-  describe("specSelectors.specOrigin", function() {
-    it("should default to 'non-editor'", function() {
+describe('editor state plugins', () => {
+  describe('specSelectors.specOrigin', () => {
+    it('should default to \'non-editor\'', () => {
       // Given
-      const system = SwaggerUi({plugins: [EditorSpecPlugin]})
+      const system = SwaggerUi({plugins: [EditorSpecPlugin]});
 
       // When
-      const res = system.specSelectors.specOrigin()
+      const res = system.specSelectors.specOrigin();
 
       // Then
-      expect(res).toEqual("not-editor")
-    })
-  })
-  describe("specActions.updateSpec", function() {
-    it("should add a parameter - origin - to state", function() {
+      expect(res).toEqual('not-editor');
+    });
+  });
+  describe('specActions.updateSpec', () => {
+    it('should add a parameter - origin - to state', () => {
       // Given
-      const system = SwaggerUi({plugins: [EditorSpecPlugin]})
+      const system = SwaggerUi({plugins: [EditorSpecPlugin]});
 
       // When
-      system.specActions.updateSpec("one: 1", "editor")
+      system.specActions.updateSpec('one: 1', 'editor');
 
       // Then
-      const res = system.specSelectors.specOrigin()
-      expect(res).toEqual("editor")
-    })
+      const res = system.specSelectors.specOrigin();
+      expect(res).toEqual('editor');
+    });
 
-    it("should default to 'non-editor'", function() {
+    it('should default to \'non-editor\'', () => {
       // Given
-      const system = SwaggerUi({plugins: [EditorSpecPlugin]})
+      const system = SwaggerUi({plugins: [EditorSpecPlugin]});
 
       // When
-      system.specActions.updateSpec("one: 1")
+      system.specActions.updateSpec('one: 1');
 
       // Then
-      const res = system.specSelectors.specOrigin()
-      expect(res).toEqual("not-editor")
-    })
-  })
-})
+      const res = system.specSelectors.specOrigin();
+      expect(res).toEqual('not-editor');
+    });
+  });
+});
