@@ -1,14 +1,14 @@
-import expect from "expect"
-import SwaggerUi from "swagger-ui"
-import debounce from "lodash/debounce"
 
-import ValidateBasePlugin from "plugins/validate-base"
-import ValidateSemanticPlugin from "plugins/validate-semantic"
-import ASTPlugin from "plugins/ast"
+import SwaggerUi from 'swagger-ui';
+import debounce from 'lodash/debounce';
 
-const envDelay = process.env.ASYNC_TEST_DELAY
+import ValidateBasePlugin from 'plugins/validate-base';
+import ValidateSemanticPlugin from 'plugins/validate-semantic';
+import ASTPlugin from 'plugins/ast';
 
-const DELAY_MS = (typeof envDelay === "string" ? parseInt(envDelay) : envDelay) || 50
+const envDelay = process.env.ASYNC_TEST_DELAY;
+
+const DELAY_MS = (typeof envDelay === 'string' ? parseInt(envDelay) : envDelay) || 50;
 
 export default function validateHelper(spec) {
   return new Promise((resolve) => {
@@ -35,38 +35,38 @@ export default function validateHelper(spec) {
               actions: {
                 loaded: () => {
                   return {
-                    type: "noop"
-                  }
+                    type: 'noop'
+                  };
                 }
               }
             }
           }
         })
       ]
-    })
+    });
 
-    system.validateActions.all()
+    system.validateActions.all();
 
-    const registerActivity = debounce(() => resolve(system), DELAY_MS)
+    const registerActivity = debounce(() => resolve(system), DELAY_MS);
 
-    system.getStore().subscribe(registerActivity)
-  })
+    system.getStore().subscribe(registerActivity);
+  });
 
 }
 
 export function expectNoErrorsOrWarnings(spec) {
   return validateHelper(spec)
     .then( system => {
-      const allErrors = system.errSelectors.allErrors().toJS()
-      expect(allErrors).toEqual([])
-    })
+      const allErrors = system.errSelectors.allErrors().toJS();
+      expect(allErrors).toEqual([]);
+    });
 }
 
 export function expectNoErrors(spec) {
   return validateHelper(spec)
     .then(system => {
-      let allErrors = system.errSelectors.allErrors().toJS()
-      allErrors = allErrors.filter(a => a.level === "error") // ignore warnings
-      expect(allErrors).toEqual([])
-    })
+      let allErrors = system.errSelectors.allErrors().toJS();
+      allErrors = allErrors.filter(a => a.level === 'error'); // ignore warnings
+      expect(allErrors).toEqual([]);
+    });
 }
