@@ -1,3 +1,2 @@
-if (typeof process === "object") {
-  require("jsdom-global")()
-}
+// Jest config globals can go here
+global.XMLHttpRequest = undefined;
