@@ -33,7 +33,16 @@ const baseRules = [
     },
   },
   { test: /\.(txt|yaml)$/, loader: "raw-loader" },
-  { test: /\.(png|jpg|jpeg|gif|svg)$/, loader: "url-loader" },
+  {
+    test: /\.(png|jpg|jpeg|gif|svg)$/, use: [
+      {
+        loader: "url-loader",
+        options: {
+          esModule: false,
+        },
+      },
+    ],
+  },
   {
     test: /\.(woff|woff2)$/,
     loader: "url-loader?",
@@ -103,7 +112,7 @@ export default function buildConfig(
             loader: "postcss-loader",
             options: {
               sourceMap: true,
-              plugins: loader => [
+              plugins: () => [
                 require("cssnano")(),
                 require("autoprefixer")(),
               ],
