@@ -2,7 +2,18 @@
  * @prettier
  */
 
+/** Dev Note: 
+ * StatsWriterPlugin is disabled by default; uncomment to enable
+ * when enabled, rebuilding the bundle will cause error for assetSizeLimit,
+ * which we want to keep out of CI/CD
+ * post build, cli command: npx webpack-bundle-analyzer <path>
+ */ 
+
 import configBuilder from "./_config-builder"
+import { DuplicatesPlugin } from "inspectpack/plugin"
+import { WebpackBundleSizeAnalyzerPlugin } from "webpack-bundle-size-analyzer"
+// import path from "path"
+// import { StatsWriterPlugin } from "webpack-stats-plugin"
 
 const result = configBuilder(
   {
@@ -14,10 +25,11 @@ const result = configBuilder(
     emitWorkerAssets: true,
   },
   {
+    mode: "production",
+
     entry: {
       "swagger-editor-bundle": [
         "./src/styles/main.less",
-        "./src/polyfills.js",
         "./src/index.js",
       ],
     },
@@ -31,6 +43,20 @@ const result = configBuilder(
       maxEntrypointSize: 1024000 * 3.25, // MB
       maxAssetSize: 1024000 * 3.25, // MB
     },
+
+    plugins: [
+      new DuplicatesPlugin({
+        // emit compilation warning or error? (Default: `false`)
+        emitErrors: false,
+        // display full duplicates information? (Default: `false`)
+        verbose: true,
+      }),
+      new WebpackBundleSizeAnalyzerPlugin("log.bundle-sizes.swagger-editor.txt"),
+      // new StatsWriterPlugin({
+      //   filename: path.join("log.bundle-stats.swagger-editor.json"),
+      //   fields: null,
+      // }),
+    ]
   }
 )
 
